Refactor: extract breathingBorder modifier
This commit extracts the `breathingBorder` modifier from `SudokuCell.kt` into a new file `AnimationModifiers.kt` within the `feature/uicore` module.

**Key Changes:**

- **New Modifier File:**
    - Created `feature/uicore/src/main/java/com/example/feature/uicore/modifiers/AnimationModifiers.kt`.
    - Moved the `breathingBorder` composable modifier function to this new file.
- **`SudokuCell.kt`:**
    - Removed the local `breathingBorder` modifier function.
    - Updated `SudokuCell` to import and use `com.example.feature.uicore.modifiers.breathingBorder`.
    - Renamed the `isFocused` parameter of the `breathingBorder` modifier to `isActive` for more general applicability.
